% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer.R
\name{eval_cancer}
\alias{eval_cancer}
\title{Evaluates accuracy and logloss for a Breast Cancer test set.}
\usage{
eval_cancer(pred)
}
\arguments{
\item{pred}{a vector with 150 numeric values representing "malignant" diagnosis probabilities for patients in the Breast Cancer
test data.}
}
\value{
a list with 2 attributes: accuracy and logloss.
}
\description{
This function evaluates the accuracy and logloss of predictions for a Breast Cancer test set (150 patients).
}
\examples{
library(isd)
pred <- runif(150)
eval_cancer(pred)
}
\author{
Olivier François
}
