% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hastib.R
\name{rhastib}
\alias{rhastib}
\title{Simulate two-class training and testing data according to Hastie and Tibishirani's model}
\usage{
rhastib(n_train = 100, n_test = 100, n_subclass = 10, sigma2 = 0.1)
}
\arguments{
\item{n_train}{an integer for the number of training samples.}

\item{n_test}{an integer for the number of test samples.}

\item{n_subclass}{an integer for the number of subclass in the Hastie and Tibishirani model.}

\item{sigma2}{a numeric for the within-group variance in the Hastie and Tibishirani model.}
}
\value{
a list with 4 attributes:

$train a matrix with n_train samples

$test  a matrix with n_test samples

$class_train a vector of class labels for the train set

$class_test  a vector of class labels for the test set
}
\description{
This function simulates training and testing data sets according to a statistical model described by Hastie and Tibshirani in their book
}
\examples{
library(isd)
X <- rhastib(n_train = 100, n_test = 100, n_subclass = 10, sigma2 = .1)
plot(X$train, col = X$class_train, pch = 19)
}
\author{
Olivier François
}
